+++ g2p_programs/models/__init__.py
@@ -17,6 +17,11 @@
 # limitations under the License.
 #
 
+from . import constants
+from . import accounting
+from . import job_related_mixin
+from . import managers
+from . import accounting
 from . import programs
 from . import registrant
 from . import registrant_attribute