+++ g2p_programs/models/program_membership.py
@@ -16,17 +16,25 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+from lxml import etree
+
 from odoo import fields, models
 
 
 class G2PProgramMembership(models.Model):
     _inherit = ["mail.thread", "mail.activity.mixin"]
+    # TODO: rename g2p.program_membership to g2p.program.beneficiaries
     _name = "g2p.program_membership"
     _description = "Program Membership"
     _order = "id desc"
 
     partner_id = fields.Many2one(
-        "res.partner", "Registrant", help="A beneficiary", required=True, tracking=True
+        "res.partner",
+        "Registrant",
+        help="A beneficiary",
+        required=True,
+        tracking=True,
+        domain=[("is_registrant", "=", True)],
     )
     program_id = fields.Many2one(
         "g2p.program", "", help="A program", required=True, tracking=True
@@ -37,10 +45,80 @@ class G2PProgramMembership(models.Model):
             ("enrolled", "Enrolled"),
             ("paused", "Paused"),
             ("exited", "Exited"),
+            ("not_eligible", "Not Eligible"),
         ],
         default="draft",
         copy=False,
+        tracking=True,
     )
+
+    enrollment_date = fields.Date(
+        "Enrollment Date", tracking=True, default=lambda self: fields.Datetime.now()
+    )
+    exit_date = fields.Date("Exit Date", tracking=True)
+
+    _sql_constraints = [
+        (
+            "program_membership_unique",
+            "unique (partner_id, program_id)",
+            "Beneficiary must be unique per program.",
+        ),
+    ]
+
+    # TODO: Implement exit reasons
+    # exit_reason_id = fields.Many2one("Exit Reason", tracking=True) Default: Completed, Opt-Out, Other
+
+    # TODO: Implement not eligible reasons
+    # Default: "Missing data", "Does not match the criterias", "Duplicate", "Other"
+    # not_eligible_reason_id = fields.Many2one("Not Eligible Reason", tracking=True)
+
+    # TODO: Add a field delivery_mechanism_id
+    # delivery_mechanism_id = fields.Many2one("Delivery mechanism type", help="Delivery mechanism")
+    # the phone number, bank account, etc.
+    delivery_mechanism_value = fields.Char("Delivery Mechanism Value", tracking=True)
+
+    # TODO: JJ - Add a field for the preferred notification method
+
+    deduplication_status = fields.Selection(
+        selection=[
+            ("new", "New"),
+            ("processing", "Processing"),
+            ("verified", "Verified"),
+            ("duplicate", "duplicate"),
+        ],
+        default="new",
+        copy=False,
+        tracking=True,
+    )
+
+    def fields_view_get(
+        self, view_id=None, view_type="form", toolbar=False, submenu=False
+    ):
+        context = self.env.context
+        result = super(G2PProgramMembership, self).fields_view_get(
+            view_id=view_id, view_type=view_type, toolbar=toolbar, submenu=submenu
+        )
+        if view_type == "form":
+            update_arch = False
+            doc = etree.XML(result["arch"])
+
+            # Check if we need to change the partner_id domain filter
+            target_type = context.get("target_type", False)
+            if target_type:
+                domain = None
+                if context.get("target_type", False) == "group":
+                    domain = "[('is_registrant', '=', True), ('is_group','=',True)]"
+                elif context.get("target_type", False) == "individual":
+                    domain = "[('is_registrant', '=', True), ('is_group','=',False)]"
+                if domain:
+                    update_arch = True
+                    nodes = doc.xpath("//field[@name='partner_id']")
+                    for node in nodes:
+                        node.set("domain", domain)
+
+            if update_arch:
+                result["arch"] = etree.tostring(doc, encoding="unicode")
+        return result
 
     def name_get(self):
         res = super(G2PProgramMembership, self).name_get()
@@ -52,3 +130,19 @@ class G2PProgramMembership(models.Model):
                 name += rec.partner_id.name
             res.append((rec.id, name))
         return res
+
+    def open_beneficiaries_form(self):
+        for rec in self:
+            return {
+                "name": "Program Beneficiaries",
+                "view_mode": "form",
+                "res_model": "g2p.program_membership",
+                "res_id": rec.id,
+                "view_id": self.env.ref("g2p_programs.view_program_membership_form").id,
+                "type": "ir.actions.act_window",
+                "target": "new",
+                "context": {
+                    "target_type": rec.program_id.target_type,
+                    "default_program_id": rec.program_id.id,
+                },
+            }