+++ g2p_registrant/models/group.py
@@ -16,8 +16,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+import logging
 
 from odoo import fields, models
+
+_logger = logging.getLogger(__name__)
 
 
 class G2PGroup(models.Model):
@@ -28,3 +31,38 @@ class G2PGroup(models.Model):
         "g2p.group.membership", "group", "Group Members"
     )
     is_partial_group = fields.Boolean("Partial Group")
+
+    def count_individuals(self, kinds=None, criteria=None):
+        self.ensure_one()
+        # Only count active groups
+
+        if self.group_membership_ids:
+            domain = [("end_date", "=?", False), ("group", "=", self.id)]
+            # To filter the membership by kinds, we first need to find the ID of the kinds we are interested in
+            if kinds is not None:
+                kind_ids = (
+                    self.env["g2p.group.membership.kind"]
+                    .search([("name", "in", kinds)])
+                    .ids
+                )
+                domain.extend([("kind", "in", kind_ids)])
+            # We can now filter the membership by this domain
+            group_membership_ids = self.group_membership_ids.search(domain).ids
+        else:
+            return 0
+
+        if len(group_membership_ids) == 0:
+            return 0
+
+        # Finally filter the res.partner that match the criteria and are related to the group
+        domain = [("individual_membership_ids", "in", group_membership_ids)]
+        domain.extend(criteria)
+        return self.env["res.partner"].search_count(domain)
+
+
+class G2PGroupKind(models.Model):
+    _name = "g2p.group.kind"
+    _description = "Group Kind"
+    _order = "id desc"
+
+    name = fields.Char("Kind")