+++ g2p_registrant/models/group_membership.py
@@ -17,7 +17,8 @@
 # limitations under the License.
 #
 
-from odoo import api, fields, models
+from odoo import _, api, fields, models
+from odoo.exceptions import ValidationError
 
 
 class G2PGroupMembership(models.Model):
@@ -40,11 +41,22 @@ class G2PGroupMembership(models.Model):
         domain=[("is_group", "=", False), ("is_registrant", "=", True)],
         tracking=True,
     )
-    kind = fields.Many2many(
-        "g2p.group.kind", string="Kind", required=True, tracking=True
+    kind = fields.Many2many("g2p.group.membership.kind", string="Kind", tracking=True)
+    start_date = fields.Datetime(
+        "Start Date", tracking=True, default=lambda self: fields.Datetime.now()
     )
-    start_date = fields.Datetime("Start Date", tracking=True)
-    end_date = fields.Datetime("End Date", tracking=True)
+    end_date = fields.Datetime(
+        "End Date", tracking=True
+    )  # TODO: Should rename `ended_date` add a check that the date is in the past
+
+    @api.constrains("individual")
+    def _check_group_members(self):
+        rec_count = 0
+        for rec in self.group.group_membership_ids:
+            if self.individual.id == rec.individual.id:
+                rec_count += 1
+        if rec_count > 1:
+            raise ValidationError(_("Duplication of Member is not allowed "))
 
     def name_get(self):
         res = super(G2PGroupMembership, self).name_get()
@@ -64,10 +76,32 @@ class G2PGroupMembership(models.Model):
             args = [("group", operator, name)] + args
         return self._search(args, limit=limit, access_rights_uid=name_get_uid)
 
+    def open_individual_form(self):
+        return {
+            "name": "Individual Member",
+            "view_mode": "form",
+            "res_model": "res.partner",
+            "res_id": self.individual.id,
+            "view_id": self.env.ref("g2p_registrant.view_individuals_form").id,
+            "type": "ir.actions.act_window",
+            "target": "new",
+        }
 
-class G2PGoupKind(models.Model):
-    _name = "g2p.group.kind"
-    _description = "Group Kind"
+    def open_group_form(self):
+        return {
+            "name": "Group Membership",
+            "view_mode": "form",
+            "res_model": "res.partner",
+            "res_id": self.group.id,
+            "view_id": self.env.ref("g2p_registrant.view_groups_form").id,
+            "type": "ir.actions.act_window",
+            "target": "new",
+        }
+
+
+class G2PGroupMembershipKind(models.Model):
+    _name = "g2p.group.membership.kind"
+    _description = "Group Membership Kind"
     _order = "id desc"
 
     name = fields.Char("Kind")