+++ g2p_registrant/models/registrant.py
@@ -25,9 +25,6 @@ class G2PRegistrant(models.Model):
 
     # Custom Fields
     address = fields.Text("Address", tracking=True)
-    addl_fields = fields.One2many(
-        "g2p.reg.attribute.value", "partner_id", "Additional Fields", tracking=True
-    )
     disabled = fields.Datetime("Date Disabled", tracking=True)
     disabled_reason = fields.Text("Reason for disabling", tracking=True)
     disabled_by = fields.Many2one("res.users", "Disabled by", tracking=True)